requirements.txt
numpy 
pandas 
scikit-learn 
matplotlib
torch
torchvideo
transformers 
pillow 
requests
